generalized verification
