+++ src/app/model/TerminalUnix.java
@@ -2,7 +2,6 @@ package app.model;
 
 import java.io.BufferedReader;
 import java.io.File;
-import java.io.IOException;
 import java.io.InputStreamReader;
 
 import javafx.scene.control.Alert;
@@ -47,32 +46,6 @@ public class TerminalUnix extends Terminal {
 		}
 
 		return output.toString();
-	}
-
-	@Override
-	public boolean canRun() throws Exception {
-		boolean canRun = false;
-		StringBuilder output = new StringBuilder();
-		try {
-			String[] compile = { "which", pdflatexPath };
-			ProcessBuilder processBuilder = new ProcessBuilder();
-			processBuilder.command(compile);
-			Process process = processBuilder.start();
-			BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-			String line;
-			while ((line = reader.readLine()) != null) {
-				output.append(line);
-			}
-
-			if (!output.toString().contains("not found")) {
-				canRun = true;
-			}
-
-		} catch (IOException e) {
-			throw new Exception("Unable to locate latexmk on your system");
-		}
-
-		return canRun;
 	}
 
 	@Override
